+++ src/libcellgui/widgets/QSFMLWidget.cpp
@@ -70,9 +70,9 @@ void QSFMLWidget::mouseReleaseEvent(QMouseEvent*)
 
 void QSFMLWidget::wheelEvent(QWheelEvent* event)
 {
-    float delta = static_cast<float>(event->angleDelta().y());
+    double delta = static_cast<double>(event->angleDelta().y());
 
-    float zoomFactor = (delta > 0) ? 0.9f : 1.1f;
+    double zoomFactor = (delta > 0) ? 0.9f : 1.1f;
     currentZoom_ *= zoomFactor;
     view_.zoom(zoomFactor);
